Factor out ConsecutiveNarrowTable
