+++ atlasdb-perf/src/main/java/com/palantir/atlasdb/performance/benchmarks/KvsGetRangeBenchmarks.java
@@ -35,7 +35,6 @@ import org.openjdk.jmh.annotations.OutputTimeUnit;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Warmup;
 
 import com.google.common.collect.Iterables;
@@ -53,146 +52,41 @@ import com.palantir.atlasdb.performance.backend.AtlasDbServicesConnector;
 import com.palantir.common.base.ClosableIterator;
 import com.palantir.util.paging.TokenBackedBasicResultsPage;
 
-@State(Scope.Benchmark)
 @BenchmarkMode(Mode.SampleTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @Warmup(iterations = 1, time = 5, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 1, time = 30, timeUnit = TimeUnit.SECONDS)
 public class KvsGetRangeBenchmarks {
+    
+    @State(Scope.Benchmark)
+    public static class CleanNarrowTable extends ConsecutiveNarrowTable {
 
-    private static final String TABLE_NAME_1 = "performance.table1";
-    private static final String ROW_COMPONENT = "key";
-    private static final String COLUMN_NAME = "value";
-    private static final byte [] COLUMN_NAME_IN_BYTES = COLUMN_NAME.getBytes(StandardCharsets.UTF_8);
-    protected static final long MIN_STORE_TS = 1L;
-    private static final long QUERY_TIMESTAMP = Long.MAX_VALUE;
-    private static final int VALUE_BYTE_ARRAY_SIZE = 100;
-    private static final long VALUE_SEED = 279L;
-    private static final int PUT_BATCH_SIZE = 1000;
-
-    public static class Table {
-        private AtlasDbServicesConnector connector;
-        private KeyValueService kvs;
-        private Random random = new Random(VALUE_SEED);
-
-        private TableReference tableRef1;
-
-        private final int numRows;
-
-        public Table(int numRows) {
-            this.numRows = numRows;
-        }
-
-        protected void setupTable(AtlasDbServicesConnector conn) {
-            this.numRows = 10000;
-            this.connector = conn;
-            this.kvs = conn.connect().getKeyValueService();
-            this.tableRef1 = Benchmarks.createTable(kvs, TABLE_NAME_1, ROW_COMPONENT, COLUMN_NAME);
-        }
-
-        public AtlasDbServicesConnector getConnector() {
-            return connector;
-        }
-
-
-        public KeyValueService getKvs() {
-            return kvs;
-        }
-
-        public Random getRandom() {
-            return random;
-        }
-
-        public TableReference getTableRef1() {
-            return tableRef1;
-        }
-
-        public int getNumRows() {
-            return numRows;
-        }
-
-        public void cleanup() throws Exception {
-            this.kvs.dropTables(Sets.newHashSet(tableRef1));
-            this.kvs.close();
-            this.connector.close();
-            this.tableRef1 = null;
-        }
-
-        protected void setupTable(AtlasDbServicesConnector conn) {
-            this.connector = conn;
-            this.kvs = conn.connect().getKeyValueService();
-            this.tableRef1 = KvsBenchmarks.createTable(kvs, TABLE_NAME_1, ROW_COMPONENT, COLUMN_NAME);
-        }
-        protected void storeData(long storeTs) {
-            for (int i = 0; i < numRows; i += PUT_BATCH_SIZE) {
-                Map<TableReference, Map<Cell, byte[]>> multiPutMap = Maps.newHashMap();
-                multiPutMap.put(tableRef1, generateBatch(i, Math.min(PUT_BATCH_SIZE, numRows - i)));
-                kvs.multiPut(multiPutMap, storeTs);
-            }
-        }
-
-        private byte[] generateValue() {
-            byte[] value = new byte[VALUE_BYTE_ARRAY_SIZE];
-            random.nextBytes(value);
-            return value;
-        }
-
-        private Map<Cell, byte[]> generateBatch(int startKey, int size) {
-            Map<Cell, byte[]> map = Maps.newHashMapWithExpectedSize(size);
-            for (int j = 0; j < size; j++) {
-                byte[] key = Ints.toByteArray(startKey + j);
-                byte[] value = generateValue();
-                map.put(Cell.create(key, COLUMN_NAME_IN_BYTES), value);
-            }
-            return map;
+        @Override
+        @Setup(Level.Trial)
+        public void setup(AtlasDbServicesConnector conn) {
+            super.setup(conn);
+            storeData();
         }
     }
 
     @State(Scope.Benchmark)
-    public static class CleanNarrowTable extends Table {
-        public CleanNarrowTable() {
-            super(10000);
-        }
+    public static class DirtyNarrowTable extends ConsecutiveNarrowTable {
 
+        @Override
         @Setup(Level.Trial)
-        public void setupData(AtlasDbServicesConnector conn) {
-            setupTable(conn);
-            storeData(MIN_STORE_TS);
-        }
-
-        @TearDown(Level.Trial)
-        public void cleanup() throws Exception {
-            super.cleanup();
-        }
-    }
-
-    @State(Scope.Benchmark)
-    public static class DirtyNarrowTable extends Table {
-        public DirtyNarrowTable() {
-            super(10000);
-        }
-
-        @Setup(Level.Trial)
-        public void setupData(AtlasDbServicesConnector conn) {
-            setupTable(conn);
-            for (long storeTs = MIN_STORE_TS; storeTs < 10; storeTs++) {
-                storeData(storeTs);
+        public void setup(AtlasDbServicesConnector conn) {
+            super.setup(conn);
+            for (long storeTs = MIN_STORE_TS; storeTs < MIN_STORE_TS + 10; storeTs++) {
+                storeData();
             }
         }
-
-        @TearDown(Level.Trial)
-        public void cleanup() throws Exception {
-            super.cleanup();
-        }
-
     }
 
-
-    protected Object getSingleRangeInner(Table table, int sliceSize) {
+    protected Object getSingleRangeInner(ConsecutiveNarrowTable table, int sliceSize) {
         RangeRequest request = Iterables.getOnlyElement(getRangeRequests(table, 1, sliceSize));
         int startRow = Ints.fromByteArray(request.getStartInclusive());
         ClosableIterator<RowResult<Value>> result =
-                table.getKvs().getRange(table.getTableRef1(), request, QUERY_TIMESTAMP);
+                table.getKvs().getRange(table.getTableRef1(), request, Long.MAX_VALUE);
         ArrayList<RowResult<Value>> list = Lists.newArrayList(result);
         KvsBenchmarks.validate(list.size() == sliceSize, "List size %s != %s", sliceSize, list.size());
         list.forEach(rowResult -> {
@@ -204,7 +98,7 @@ public class KvsGetRangeBenchmarks {
         return result;
     }
 
-    private Iterable<RangeRequest> getRangeRequests(Table table, int numRequests, int sliceSize) {
+    private Iterable<RangeRequest> getRangeRequests(ConsecutiveNarrowTable table, int numRequests, int sliceSize) {
         List<RangeRequest> requests = Lists.newArrayList();
         Set<Integer> used = Sets.newHashSet();
         for (int i = 0; i < numRequests; i++) {
@@ -225,10 +119,10 @@ public class KvsGetRangeBenchmarks {
     }
 
 
-    protected Object getMultiRangeInner(Table table) {
+    protected Object getMultiRangeInner(ConsecutiveNarrowTable table) {
         Iterable<RangeRequest> requests = getRangeRequests(table, (int) (table.getNumRows() * 0.1), 1);
         Map<RangeRequest, TokenBackedBasicResultsPage<RowResult<Value>, byte[]>> results =
-                table.getKvs().getFirstBatchForRanges(table.getTableRef1(), requests, QUERY_TIMESTAMP);
+                table.getKvs().getFirstBatchForRanges(table.getTableRef1(), requests, Long.MAX_VALUE);
 
         int numRequests = Iterables.size(requests);
